Move functions to modules file
