+++ src/modules/reservation.js
@@ -1,5 +1,7 @@
 import { invAPI } from './api.js';
 
+/* API Functions */ 
+
 const reservationURL = `${invAPI}reservations?item_id=`;
 
 const getReservation = async (id) => {
@@ -8,12 +10,6 @@ const getReservation = async (id) => {
   console.log(data);
   return data;
 };
-
-// const displayReservation = async () => {
-//   const reservationList = await getReservation();
-//   reservationList.forEach((item) => {
-//   });
-// };
 
 const addReservation = async (id, name, dateStart, dateEnd) => {
   const res = await fetch(reservationURL + id, {
@@ -26,5 +22,46 @@ const addReservation = async (id, name, dateStart, dateEnd) => {
   const data = await res.json();
   return data;
 };
+
+/* Display functions */
+
+const newReservation = document.querySelector('#add-reservation');
+const reservationForm = document.querySelector('.reservation-form');
+const reserveBtn = document.querySelector('.reserve-button');
+const reservationPopup = document.querySelector('#reservation-popup');
+const closeBtn = document.querySelector('#close-btn');
+const list = document.querySelector('#reservation-list');
+
+const displayReservation = async () => {
+  const reservationList = await getReservation('item2');
+  reservationList.forEach((item) => {
+    list.innerHTML += `<li class="reservationName"> ${item.username}: ${item.date_start} ${item.date_end}`;
+  });
+};
+
+reserveBtn.addEventListener('click', () => {
+  reservationPopup.classList.remove('reservation-overlay');
+});
+
+closeBtn.addEventListener('click', () => {
+  reservationPopup.classList.add('reservation-overlay');
+});
+
+newReservation.addEventListener('click', (e) => {
+  e.preventDefault;
+  const reserverName = document.querySelector('#reservationName').value;
+  const startDate = document.querySelector('#startDate').value;
+  const endDate = document.querySelector('#endDate').value;
+  getReservation('item2');
+  if (!reserverName || !startDate || !endDate) {
+    e.preventDefault();
+    console.log('oops!');
+  } else {
+    addReservation('item2', reserverName, startDate, endDate);
+    //reservation-popup.innerHTML = '';
+    reservationForm.reset();
+  }
+});
+
 
 export { getReservation, addReservation };
